refactor: extract DateBadgeChips from ItemRow and DateItemRow
