+++ crates/frontend/src/components/items/mod.rs
@@ -1,5 +1,6 @@
 pub mod add_input;
 pub mod add_item_input;
+pub mod date_badge_chips;
 pub mod date_editor;
 pub mod date_item_row;
 pub mod inline_date_editor_section;
